Answer the review: restore a lost TODO, split the helper, fix the fixtures
- Restore the lost-answer TODO my edit truncated away. The replacement ran
  to end-of-file and took an unrelated, still-open item with it.
- Move the page-filling seeder into its own module; attendees.ts was over
  the 400-line limit at 417.
- Use real listing types in the new fixtures; "one_off" is not one the
  schema permits.
- Make the overbook test push the listing OVER capacity. At exactly full a
  quantity-0 row still fits, so the earlier fixture proved nothing — the
  mutant is now killed.
